funcao de remover tuplas

+++ GerenciadorArtigo.java
@@ -1,6 +1,10 @@
 // Event Listener
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import javax.swing.JOptionPane;
+import java.sql.SQLException;
+
+
 
 public class GerenciadorArtigo extends Gerenciador{
 
@@ -36,17 +40,29 @@ public class GerenciadorArtigo extends Gerenciador{
 			// Seleciona linha selecionada pelo usuario
 			int linhaSelecionada =  table.getSelectedRow();
 
-			// Seleciona ID do Artigo selecionado, chave primaria para remocao
-			String removerId = dados[linhaSelecionada][0];
+			if(linhaSelecionada != -1){
 
-			String query = "DELETE FROM ARTIGO WHERE idArt = " + removerId;
-			System.out.println(query);
 
-			// Remove da tabela o artigo
-			this.dbcon.executarQuery(query);
+				// Seleciona ID do Artigo selecionado, chave primaria para remocao
+				String removerId = dados[linhaSelecionada][0];
 
-			removerLinha(linhaSelecionada);
+				String query = "DELETE FROM ARTIGO WHERE idArt = " + removerId;
+				System.out.println(query);
 
+				// Remove da tabela o artigo
+				try{
+					this.dbcon.executarQuery(query);
+					removerLinha(linhaSelecionada);
+				}
+				catch(SQLException ex){
+					JOptionPane.showMessageDialog(null, ex.getMessage(), "Erro", JOptionPane.ERROR_MESSAGE);
+
+				}
+			}
+			else{
+				JOptionPane.showMessageDialog(null, "Nenhuma linha selecionada", "Erro", JOptionPane.ERROR_MESSAGE);
+			}
+			
 		}
 		else if(e.getActionCommand().equals(editar.getText()))
 		{
